[Silver V] Title: 수 이어가기, Time: 0 ms, Memory: 2020 KB -BaekjoonHub

+++ 백준/Silver/2635. 수 이어가기/수 이어가기.cc
@@ -26,7 +26,7 @@ int main()
     }
 
     cout << maxLen << '\n';
-    for (size_t i = 0; i < maxArr.size(); i++) {
+    for (int i = 0; i < maxArr.size(); i++) {
         cout << maxArr[i];
         if (i != maxArr.size() - 1) cout << ' ';
     }
